# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing loans with a state machine workflow",
        "title": "Loan Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5002",
    "basePath": "/api/v1",
    "paths": {
...
        "/loans": {
            "get": {
                "description": "Get a list of all loans with optional filtering",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "List all loans",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Maximum loan amount",
                        "name": "max_amount",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum loan amount",
                        "name": "min_amount",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new loan with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Create a new loan",
                "parameters": [
                    {
                        "description": "Loan information",
                        "name": "loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/loans/{id}/{status}": {
            "patch": {
                "description": "Update a loan's status based on the provided status transition\n- For approve: { \"success\": true, \"message\": \"Loan status updated successfully\" }\n- For partial invest: { \"success\": true, \"data\": { \"remaining_amount\": 150000, \"invested_amount\": 50000, \"agreement_document\": null }, \"message\": \"loan invested successfully\" }\n- For full invest: { \"success\": true, \"data\": { \"remaining_amount\": 0, \"invested_amount\": 200000, \"agreement_document\": \"agreement_file.pdf\" }, \"message\": \"loan status updated to invested\" }\n- For disburse: { \"success\": true, \"data\": { \"field_officer_id\": \"emp-789\", \"agreement_file_name\": \"agreement.pdf\" }, \"message\": \"loan disbursed successfully\" }",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Update loan status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New status",
                        "name": "status",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status update information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    },
                    {
                        "description": "Approve request (when status=approve)",
                        "name": "approveRequest",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/swagger.ApproveSchema"
                        }
                    },
                    {
                        "description": "Invest request (when status=invest)",
                        "name": "investRequest",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/swagger.InvestSchema"
                        }
                    },
                    {
                        "description": "Disburse request (when status=disburse)",
                        "name": "disburseRequest",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/swagger.DisburseSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful status update with varying response structure based on status",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or status transition",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "pagination": {
                    "$ref": "#/definitions/domain.PaginationInfo"
                }
            }
        },
        "domain.PaginationInfo": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "request.CreateLoanRequest": {
            "type": "object",
            "required": [
                "amount",
                "borrowerId",
                "rate",
                "roi"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "borrowerId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "rate": {
                    "type": "number"
                },
                "roi": {
                    "type": "number"
                }
            }
        },
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "swagger.ApproveSchema": {
            "type": "object",
            "properties": {
                "approval_date": {
                    "type": "string",
                    "example": "2025-03-25"
                },
                "approval_employee_id": {
                    "type": "string",
                    "example": "emp-123"
                },
                "file_name": {
                    "type": "string",
                    "example": "approval_document.pdf"
                }
            }
        },
        "swagger.DisburseSchema": {
            "type": "object",
            "properties": {
                "agreement_file_name": {
                    "type": "string",
                    "example": "loan_agreement.pdf"
                },
                "field_officer_id": {
                    "type": "string",
                    "example": "emp-789"
                }
            }
        },
        "swagger.InvestSchema": {
            "type": "object",
            "properties": {
                "invest_amount": {
                    "type": "number",
                    "example": 5000
                },
                "lender_id": {
                    "type": "string",
                    "example": "lender-456"
                }
            }
        }
    }
}